requirements.txt
SQLAlchemy~=1.4.36
python-dotenv~=0.20.0
pydantic~=1.9.0
uvicorn~=0.17.6
fastapi~=0.75.2
requests~=2.27.1
alembic~=1.7.7
setuptools~=62.1.0
psycopg2-binary==2.9.3
